ModSprite Separation.
